Trait ethcore_ipc::interface::IpcConfig

pub trait IpcConfig {
    fn api_version() -> Version { ... }
    fn protocol_version() -> Version { ... }
    fn handshake(handshake: &Handshake) -> bool { ... }
}

Allows to configure custom version and custom handshake response for ipc host

Provided Methods

fn api_version() -> Version

Current service api version Should be increased if any of the methods changes signature

fn protocol_version() -> Version

Current ipc protocol version Should be increased only if signature of system methods changes

fn handshake(handshake: &Handshake) -> bool

Default handshake requires exact versions match